Consolidate Token example code across webapp components [M] (#346)
