+++ webapp/src/components/CodeComparison.tsx
@@ -1,3 +1,4 @@
+import { EXAMPLES, COMPARISON_SOL_LINES } from "../examples.ts";
 import WindowDots from "./WindowDots.tsx";
 import SyntaxLine from "./SyntaxLine.tsx";
 
@@ -27,61 +28,7 @@ function CodeComparisonPanel({
   );
 }
 
-const compTsLines = [
-  'import { address, msg } from "skittles";',
-  "",
-  "export class Token {",
-  "  totalSupply: number = 0;",
-  "  private balances: Record<address, number> = {};",
-  "",
-  "  constructor(supply: number) {",
-  "    this.totalSupply = supply;",
-  "    this.balances[msg.sender] = supply;",
-  "  }",
-  "",
-  "  balanceOf(account: address): number {",
-  "    return this.balances[account];",
-  "  }",
-  "",
-  "  transfer(to: address, amount: number): boolean {",
-  "    if (this.balances[msg.sender] < amount) {",
-  '      throw new Error("Insufficient balance");',
-  "    }",
-  "    this.balances[msg.sender] -= amount;",
-  "    this.balances[to] += amount;",
-  "    return true;",
-  "  }",
-  "}",
-];
-
-const compSolLines = [
-  "// SPDX-License-Identifier: MIT",
-  "pragma solidity ^0.8.20;",
-  "",
-  "contract Token {",
-  "    uint256 public totalSupply;",
-  "    mapping(address => uint256) internal balances;",
-  "",
-  "    constructor(uint256 supply) {",
-  "        totalSupply = supply;",
-  "        balances[msg.sender] = supply;",
-  "    }",
-  "",
-  "    function balanceOf(address account)",
-  "        public view virtual returns (uint256) {",
-  "        return balances[account];",
-  "    }",
-  "",
-  "    function transfer(address to, uint256 amount)",
-  "        public virtual returns (bool) {",
-  "        require(balances[msg.sender] >= amount,",
-  '            "Insufficient balance");',
-  "        balances[msg.sender] -= amount;",
-  "        balances[to] += amount;",
-  "        return true;",
-  "    }",
-  "}",
-];
+const compTsLines = EXAMPLES.Token.trimEnd().split("\n");
 
 function CodeComparison() {
   return (
@@ -93,7 +40,7 @@ function CodeComparison() {
       </p>
       <div className="comparison-row">
         <CodeComparisonPanel label="Token.ts" lines={compTsLines} lang="ts" />
-        <CodeComparisonPanel label="Token.sol — generated" lines={compSolLines} lang="sol" highlighted />
+        <CodeComparisonPanel label="Token.sol — generated" lines={COMPARISON_SOL_LINES} lang="sol" highlighted />
       </div>
       <a href="#playground" className="btn-primary btn-primary--lg comparison-cta">
         <span>Open the Playground</span>
